Fixes permissions not saving
Fixes an issue in newer versions of DNN where the module permissions (submit/feature/meta/etc) would not save if an entire column was empty. It seems in newer versions of DNN, if a module setting is empty, it simply doesn't create a row in the database. This means that the setting is set to null which means all are checked, making it impossible for a column to be totally unchecked.